separate Padding utils and add more comments
